requirements.txt
fastapi
uvicorn
requests
pandas
numpy
python-dotenv
groq
resend
setuptools